myers: Split code into multiple files
