+++ src/pattern_matching/myers/mod.rs
@@ -189,6 +189,12 @@ use num_traits::{Bounded, FromPrimitive, One, PrimInt, ToPrimitive, WrappingAdd,
 use alignment::{Alignment, AlignmentMode, AlignmentOperation};
 use utils::{IntoTextIterator, TextIterator};
 
+mod builder;
+mod traceback;
+
+pub use self::builder::MyersBuilder;
+use self::traceback::Traceback;
+
 /// Integer types serving as bit vectors must implement this trait.
 /// Only unsigned integers will work correctly.
 pub trait BitVec: Copy
@@ -239,184 +245,6 @@ pub type Myers64 = Myers<u64>;
 /// Myers instance using `u128` as bit vectors (pattern length up to 128)
 #[cfg(has_u128)]
 pub type Myers128 = Myers<u128>;
-
-/// Builds a Myers instance, allowing to specify ambiguities.
-///
-/// # Example:
-///
-/// This example shows how recognition of IUPAC ambiguities in patterns
-/// can be implemented. Ambiguities in the searched text are not recognized; this
-/// would require specifying additional ambiguities (A = MRWVHDN, etc...).
-///
-/// ```
-/// # extern crate bio;
-/// use bio::pattern_matching::myers::MyersBuilder;
-///
-/// # fn main() {
-/// let ambigs = [
-///     (b'M', &b"ACM"[..]),
-///     (b'R', &b"AGR"[..]),
-///     (b'W', &b"ATW"[..]),
-///     (b'S', &b"CGS"[..]),
-///     (b'Y', &b"CTY"[..]),
-///     (b'K', &b"GTK"[..]),
-///     (b'V', &b"ACGMRSV"[..]),
-///     (b'H', &b"ACTMWYH"[..]),
-///     (b'D', &b"AGTRWKD"[..]),
-///     (b'B', &b"CGTSYKB"[..]),
-///     (b'N', &b"ACGTMRWSYKVHDBN"[..])
-/// ];
-///
-/// let mut builder = MyersBuilder::new();
-///
-/// for &(base, equivalents) in &ambigs {
-///     builder.ambig(base, equivalents);
-/// }
-///
-/// let text = b"GGATGNGCGCCATAG";
-/// let pattern = b"TRANCGG";
-/// //                *   * (mismatch)
-///
-/// let myers = builder.build(pattern);
-/// assert_eq!(myers.distance(text), 2);
-/// # }
-/// ```
-#[derive(Default, Clone, Eq, PartialEq)]
-pub struct MyersBuilder {
-    ambigs: HashMap<u8, Vec<u8>>,
-    wildcards: Vec<u8>,
-}
-
-impl MyersBuilder {
-    pub fn new() -> MyersBuilder {
-        Self::default()
-    }
-
-    /// Allows to specify ambiguous characters and their equivalents.
-    ///
-    /// # Example:
-    ///
-    /// ```
-    /// # extern crate bio;
-    /// use bio::pattern_matching::myers::MyersBuilder;
-    ///
-    /// # fn main() {
-    /// let text = b"ACCGTGGATGAGCGCCATAG";
-    /// let pattern =      b"TGAGCGN";
-    ///
-    /// let myers = MyersBuilder::new()
-    ///     .ambig(b'N', b"ACGT")
-    ///     .build(pattern);
-    ///
-    /// assert_eq!(myers.distance(text), 0);
-    /// # }
-    pub fn ambig<I, B>(&mut self, byte: u8, equivalents: I) -> &mut Self
-    where
-        I: IntoIterator<Item = B>,
-        B: Borrow<u8>,
-    {
-        let eq = equivalents.into_iter().map(|b| *b.borrow()).collect();
-        self.ambigs.insert(byte, eq);
-        self
-    }
-
-    /// Allows to specify a wildcard character, that upon appearance in the search text
-    /// shall be matched by any character of the pattern. Multiple wildcards are possible.
-    /// For the inverse, that is, wildcards in the pattern matching any character in search
-    /// text, use `ambig(byte, 0..255)`.
-    ///
-    /// # Example:
-    ///
-    /// ```
-    /// # extern crate bio;
-    /// use bio::pattern_matching::myers::MyersBuilder;
-    ///
-    /// # fn main() {
-    /// let text = b"ACCGTGGATGAGCG*CATAG";
-    /// let pattern =      b"TGAGCGT";
-    ///
-    /// let myers = MyersBuilder::new()
-    ///     .text_wildcard(b'*')
-    ///     .build(pattern);
-    ///
-    /// assert_eq!(myers.distance(text), 0);
-    /// # }
-    pub fn text_wildcard(&mut self, wildcard: u8) -> &mut Self {
-        self.wildcards.push(wildcard);
-        self
-    }
-
-    /// Creates a Myers instance given a pattern, using `u64` as bit vector type
-    pub fn build<'a, P>(&self, pattern: P) -> Myers<u64>
-    where
-        P: IntoTextIterator<'a>,
-        P::IntoIter: ExactSizeIterator,
-    {
-        self.build_other(pattern)
-    }
-
-    /// Creates a Myers instance given a pattern, using `u128` as bit vector type
-    #[cfg(has_u128)]
-    pub fn build128<'a, P>(&self, pattern: P) -> Myers<u128>
-    where
-        P: IntoTextIterator<'a>,
-        P::IntoIter: ExactSizeIterator,
-    {
-        self.build_other(pattern)
-    }
-
-    /// Creates a Myers instance given a pattern, using any desired type for bit vectors
-    ///
-    /// # Example:
-    ///
-    /// ```
-    /// # extern crate bio;
-    /// use bio::pattern_matching::myers::{MyersBuilder, Myers};
-    ///
-    /// # fn main() {
-    /// let myers: Myers<u32> = MyersBuilder::new()
-    ///     .text_wildcard(b'*')
-    ///     .build_other(b"TGAGCG*");
-    /// // ...
-    /// # }
-    pub fn build_other<'a, T, P>(&self, pattern: P) -> Myers<T>
-    where
-        T: BitVec,
-        P: IntoTextIterator<'a>,
-        P::IntoIter: ExactSizeIterator,
-    {
-        let maxsize = T::DistType::from_usize(size_of::<T>() * 8).unwrap();
-        let pattern = pattern.into_iter();
-        let m = T::DistType::from_usize(pattern.len()).unwrap();
-        assert!(m <= maxsize, "Pattern too long");
-        assert!(m > T::DistType::zero(), "Pattern is empty");
-
-        let mut peq = [T::zero(); 256];
-
-        for (i, &a) in pattern.enumerate() {
-            let mask = T::one() << i;
-            // equivalent
-            peq[a as usize] |= mask;
-            // ambiguities
-            if let Some(equivalents) = self.ambigs.get(&a) {
-                for &eq in equivalents {
-                    peq[eq as usize] |= mask;
-                }
-            }
-        }
-
-        for &w in &self.wildcards {
-            peq[w as usize] = T::max_value();
-        }
-
-        Myers {
-            peq: peq,
-            bound: T::one() << (m.to_usize().unwrap() - 1),
-            m: m,
-            tb: Traceback::new(),
-        }
-    }
-}
 
 /// Myers algorithm.
 pub struct Myers<T>
@@ -867,275 +695,6 @@ fn update_aln(
     aln.ystart = aln.yend - aln_len;
     aln.mode = AlignmentMode::Semiglobal;
     aln.score = dist as i32;
-}
-
-struct Traceback<T: BitVec> {
-    states: Vec<State<T>>,
-    positions: iter::Cycle<Range<usize>>,
-    pos: usize,
-    m: T::DistType,
-}
-
-impl<T> Traceback<T>
-where
-    T: BitVec,
-{
-    fn new() -> Traceback<T> {
-        Traceback {
-            states: vec![],
-            positions: (0..0).cycle(),
-            m: T::DistType::zero(),
-            pos: 0,
-        }
-    }
-
-    fn init(&mut self, initial_state: State<T>, num_cols: usize, m: T::DistType) {
-        // ensure that empty text does not cause panic
-        let num_cols = num_cols + 2;
-
-        self.positions = (0..num_cols).cycle();
-
-        // extend or truncate states vector
-        let curr_len = self.states.len();
-        if num_cols > curr_len {
-            self.states.reserve(num_cols);
-            self.states
-                .extend((0..num_cols - curr_len).map(|_| State::default()));
-        } else {
-            self.states.truncate(num_cols);
-        }
-        // important if using unsafe in add_state(), and also for correct functioning of traceback
-        debug_assert!(self.states.len() == num_cols);
-
-        // first column is used to ensure a correct path if the text
-        // is shorter than the pattern
-        // Actual leftmost column starts at second position
-        self.add_state(State {
-            dist: T::DistType::max_value(),
-            pv: T::max_value(), // all 1s
-            mv: T::zero(),
-        });
-
-        // initial state
-        self.add_state(initial_state);
-
-        self.m = m;
-    }
-
-    #[inline]
-    fn add_state(&mut self, s: State<T>) {
-        self.pos = self.positions.next().unwrap();
-        //self.states[self.pos] = s;
-        // faster and safe since the positions are cycled.
-        // Unless the code in init() is wrong, the index should
-        // never be out of bounds.
-        *unsafe { self.states.get_unchecked_mut(self.pos) } = s;
-    }
-
-    // Returns the length of the current match, optionally adding the
-    // alignment path to `ops`
-    fn traceback(&self, ops: Option<&mut Vec<AlignmentOperation>>) -> (usize, T::DistType) {
-        self._traceback_at(self.pos, ops)
-    }
-
-    // Returns the length of a match with a given end position, optionally adding the
-    // alignment path to `ops`
-    // only to be called if the `states` vec contains all states of the text
-    fn traceback_at(
-        &self,
-        pos: usize,
-        ops: Option<&mut Vec<AlignmentOperation>>,
-    ) -> Option<(usize, T::DistType)> {
-        let pos = pos + 2; // in order to be comparable since self.pos starts at 2, not 0
-        if pos <= self.pos {
-            return Some(self._traceback_at(pos, ops));
-        }
-        None
-    }
-
-    // returns a tuple of alignment length and hit distance, optionally adding the alignment path
-    // to `ops`
-    #[inline]
-    fn _traceback_at(
-        &self,
-        pos: usize,
-        mut ops: Option<&mut Vec<AlignmentOperation>>,
-    ) -> (usize, T::DistType) {
-        use self::AlignmentOperation::*;
-
-        // Reverse iterator over states. If remembering all positions,
-        // the chain() and cycle() are not actually needed, but there seems
-        // to be almost no performance loss.
-        let mut states = self.states[..pos + 1]
-            .iter()
-            .rev()
-            .chain(self.states.iter().rev().cycle());
-        // // Simpler alternative using skip() is slower in some cases:
-        // let mut states = self.states.iter().rev().cycle().skip(self.states.len() - pos - 1);
-
-        // self.print_tb_matrix(pos);
-
-        let ops = &mut ops;
-        if let Some(o) = ops.as_mut() {
-            o.clear();
-        }
-
-        // Mask for testing current position. The mask is moved
-        // along mv / pv for testing all positions
-        let mut current_pos = T::one() << (self.m.to_usize().unwrap() - 1);
-
-        // horizontal column offset from starting point of traceback matrix (bottom right)
-        let mut h_offset = 0;
-        // vertical column offset from starting point of traceback matrix (bottom right)
-        let mut v_offset = 0;
-
-        // Macro for moving up one cell in traceback matrix, adjusting the distance of the state
-        // of the state. Expects a bit mask indicating the current row position in the matrix.
-        macro_rules! move_state_up {
-            ($state:expr, $pos_mask:expr) => {
-                if $state.pv & $pos_mask != T::zero() {
-                    $state.dist -= T::DistType::one()
-                } else if $state.mv & $pos_mask != T::zero() {
-                    $state.dist += T::DistType::one()
-                }
-            };
-        }
-
-        // Macro for moving up 'n' cells in traceback matrix, adjusting the distance of the state.
-        macro_rules! move_up_many {
-            ($state:expr, $n:expr) => {
-                // construct mask covering bits to check
-                let m = self.m.to_usize().unwrap();
-                let mask = if $n.to_usize().unwrap() == size_of::<T>() * 8 {
-                    T::max_value()
-                } else {
-                    (!(T::max_value() << $n)) << (m - $n)
-                };
-                // adjust distance
-                let p = ($state.pv & mask).count_ones() as i32;
-                let m = ($state.mv & mask).count_ones() as i32;
-                $state.dist = T::DistType::from_i32($state.dist.to_i32().unwrap() - p + m).unwrap();
-
-                // // A loop seems always slower (not sure about systems without popcnt):
-                // let mut p =  T::one() << (self.m.to_usize().unwrap() - 1);
-                // for _ in 0..$n {
-                //     move_state_up!($state, p);
-                //     p >>= 1;
-                // }
-            };
-        }
-
-        // current state
-        let mut state = states.next().unwrap().clone();
-        // distance of the match
-        let dist = state.dist;
-        // state left to the current state
-        let mut lstate = states.next().unwrap().clone();
-        // The distance of the left state is always for diagonal of to the current position
-        // in the traceback matrix. This allows checking for a substitution by a simple
-        // comparison.
-        move_state_up!(lstate, current_pos);
-
-        // Macro for collectively moving up the cursor (adjusting the distance) of both
-        // the left and the current state in the traceback matrix
-        macro_rules! move_up {
-            () => {
-                v_offset += 1;
-                move_state_up!(state, current_pos);
-                current_pos >>= 1;
-                move_state_up!(lstate, current_pos);
-            };
-        }
-
-        // Macro for moving to the left, adjusting h_offset
-        macro_rules! move_left {
-            () => {
-                h_offset += 1;
-                state = lstate;
-                lstate = states.next().unwrap().clone();
-                if v_offset != self.m.to_usize().unwrap() {
-                    // move up v_offset + 1 (not only v_offset) because left state
-                    // is always in diagonal position
-                    move_up_many!(lstate, v_offset + 1);
-                }
-            };
-        }
-
-        // Loop for finding the traceback path
-        while v_offset < self.m.to_usize().unwrap() {
-            let op = if lstate.dist + T::DistType::one() == state.dist {
-                // Diagonal (substitution)
-                // Since cursor of left state is always in the diagonal position,
-                // this is a simple comparison of distances.
-                v_offset += 1;
-                current_pos >>= 1;
-                move_left!();
-                Subst
-            } else if state.pv & current_pos != T::zero() {
-                // Up
-                move_up!();
-                Ins
-            } else if lstate.mv & current_pos != T::zero() {
-                // Left
-                move_left!();
-                state.dist -= T::DistType::one();
-                Del
-            } else {
-                // Diagonal (match)
-                v_offset += 1;
-                current_pos >>= 1;
-                move_left!();
-                Match
-            };
-
-            if let Some(o) = ops.as_mut() {
-                o.push(op);
-            }
-        }
-
-        if let Some(o) = ops.as_mut() {
-            o.reverse();
-        }
-
-        (h_offset, dist)
-    }
-
-    // Useful for debugging
-    #[allow(dead_code)]
-    fn print_tb_matrix(&self, pos: usize) {
-        let states = self.states[..pos + 1]
-            .iter()
-            .rev()
-            .chain(self.states.iter().rev().cycle());
-
-        let m = self.m.to_usize().unwrap();
-
-        let mut out: Vec<_> = (0..m + 1).map(|_| vec![]).collect();
-        for mut s in states.into_iter().cloned() {
-            let mut current_pos = T::one() << (m - 1);
-            let mut dist = s.dist;
-            for i in 0..m + 1 {
-                out[i].push(dist.to_usize().unwrap());
-                if s.pv & current_pos != T::zero() {
-                    dist -= T::DistType::one();
-                } else if s.mv & current_pos != T::zero() {
-                    dist += T::DistType::one();
-                }
-                current_pos >>= 1;
-            }
-            if s.dist == T::DistType::max_value() {
-                break;
-            }
-        }
-        out.reverse();
-        for mut line in out {
-            line.reverse();
-            for dist in line {
-                print!("{:<4}", dist);
-            }
-            print!("\n");
-        }
-    }
 }
 
 // temporary impl to create new 'empty' alignment
